Refactor and fix naming issue
