+++ .coveragerc
@@ -14,7 +14,7 @@ omit =
     # Skip Dask until it is fully a part of the testing suite
     modin/engines/dask/*
     # Skip Gandiva because it is experimental
-    modin/engines/ray/gandiva_on_ray/*
+    modin/experimental/engines/ray/gandiva_on_ray/*
     modin/data_management/query_compiler/gandiva_query_compiler.py
 
 [report]
